pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.121.0",
    "openai-agents>=0.3.3",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.37.0",
]
